requirements.txt
argcomplete==1.10.0
beautifulsoup4==4.8.0
chardet==3.0.4
click==7.1.2
Django==2.2.14
django-cors-headers==3.4.0
djangorestframework==3.11.0
docx2txt==0.8
EbookLib==0.17.1
extract-msg==0.23.1
IMAPClient==2.1.0
joblib==0.14.1
lxml==4.5.2
nltk==3.5
numpy==1.18.5
olefile==0.46
pandas==0.24.2
pdfminer.six==20181108
Pillow==7.2.0
pkg-resources==0.0.0
pycryptodome==3.9.8
PyPDF2==1.26.0
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2020.1
regex==2020.6.8
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.12.0
sklearn==0.0
sortedcontainers==2.2.2
soupsieve==2.0.1
SpeechRecognition==3.8.1
sqlparse==0.3.1
textblob==0.15.3
textract==1.6.3
tqdm==4.47.0
tzlocal==1.5.1
xlrd==1.2.0
XlsxWriter==1.2.9
